Rm unused files

+++ include/parser.h
@@ -1,8 +1,7 @@
 #ifndef PARSER_H
 # define PARSER_H
-#include "settings.h"
 #include "elements.h"
-#include "minirt.h"
+#include "miniRT.h"
 
 //parser.c
 int	pars(const char *path, t_app *app);
@@ -30,6 +29,6 @@ t_vec3	extract_normal(const char *input);
 //free_tokens.c
 void	free_tokens(char **tokens);
 int	token_ammount(char **tokens);
-FLOAT	ft_atof(const char *input);
+double	ft_atof(const char *input);
 
 #endif
